backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.10.1
python-multipart==0.0.20
PyYAML==6.0.2
sentence-transformers==5.1.0
numpy==2.3.2
email-validator==2.2.0
httpx==0.28.1
pytest==8.4.1
